+++ test/clojure2d/color_test.clj
@@ -249,4 +249,3 @@
 ;; test paletton generator
 
 ;; test nearest color distance
-